refactor(frontend): rename "staking" to "stake"
