+++ src/frontend/src/icp/components/stake/gldt/GldtStake.svelte
